feat(workspace): Workspace shell + home / bar カラム (マルチカラム PR 2/5)
`/` の index route を単一ホーム表示から Workspace (マルチカラム) に切替。
docs/16-multicolumn.md 予定の「workspace as default route」の最初の実体。

components/workspace.tsx (新規):
- loadAppColumns() で読んだ構成を ColumnView の並びとして表示
- ColumnView は header + 縦スクロールする body。body 要素は
  useState + callback ref で持ち、ColumnScrollContext で配下に配る
  (= VirtualFeed がカラム内スクロールへ自然に切り替わる、PR 1 の
  element prop 設計に対応)
- columns は useMemo で安定化 (default 構成は呼ぶたび新 id を生成する
  ため、毎 render 再計算すると全カラム remount になる)
- session loading 中は構成を読まない (false で読むと board 構成が
  1 render チラつくため)。未サインインは旧 home.tsx の landing を表示

column-content/ (新規):
- index.tsx: kind 別 dispatcher。未実装 kind (notifications / search /
  board / profile) は従来ページへのリンク付きプレースホルダ
  (flag なし段階導入のため dev でも壊れた見た目にならない)
- home-column.tsx: 旧 home.tsx の「フォロー」タブを移設。HomeSummary +
  投稿ボタン + フォロー TL (useInfiniteFeed + IDB cache)。tab 条件が
  消えたぶん enabled / deps が簡素化
- bar-column.tsx: 旧「共鳴」タブを BAR ブルスコ (酒場メタファー) として
  独立。冒頭に 1 文の説明を追加

routes/home.tsx は削除 (全機能が workspace + column-content に移設)。

styles.css:
- .workspace-columns: モバイル縦積み / 768px+ 横並び (横スワイプ snap は
  PR 5)
- .workspace-column: dq-window 様式 (3px 白枠 + 四隅装飾) の独立
  スクロールカラム。768px+ は固定幅 340px + 高さ calc(100dvh - 9.5em)
  (実測 CSS 変数化は PR 5)
- workspace モードでは .content の大枠 (二重枠) を透明化
  (= カラム自体が手帳 1 ページ、workspace は見開きのメタファー)
- 1100px+ で app-shell の max-width を 1480px まで拡張

検証: typecheck / vitest 119 件 / build 緑。playwright smoke で
未サインイン landing 表示 + /board /search /onboarding の無事 +
console エラーなしを確認 (サインイン後のカラム表示は dev 環境で
実機確認が必要)。
