+++ apps/web/src/styles.css
@@ -205,6 +205,95 @@ p { margin: 0.4em 0; }
   }
 }
 
+/* ─── アプリ全体マルチカラム workspace (docs/16-multicolumn.md) ── */
+/* `/` の index route が [data-workspace] を含むとき、app-shell の幅制限を
+   緩める。各カラムが DQ ウィンドウになるので、.content の大枠 (二重枠) は
+   workspace モードでは透明化する (= 「手帳の見開き」メタファー)。 */
+@media (min-width: 1100px) {
+  .app-shell:has([data-workspace="1"]) { max-width: min(100vw - 2em, 1480px); }
+}
+
+.app-shell:has([data-workspace="1"]) .content {
+  background: transparent;
+  border-color: transparent;
+  box-shadow: none;
+  padding: 0.4em 0.2em;
+}
+.app-shell:has([data-workspace="1"]) .content::before,
+.app-shell:has([data-workspace="1"]) .content::after {
+  display: none;
+}
+
+.workspace-columns {
+  display: flex;
+  flex-direction: column;     /* モバイル: 縦積み (横スワイプ snap は PR 5) */
+  gap: 0.7em;
+}
+
+@media (min-width: 768px) {
+  .workspace-columns {
+    flex-direction: row;
+    gap: 0.8em;
+    overflow-x: auto;
+    align-items: flex-start;
+    padding-bottom: 0.5em;
+  }
+}
+
+.workspace-column {
+  display: flex;
+  flex-direction: column;
+  background: var(--color-window-bg);
+  border: 3px solid var(--color-border);
+  border-radius: 3px;
+  box-shadow:
+    inset 0 0 0 1px var(--color-window-inner-border),
+    0 4px 0 rgba(0, 0, 0, 0.35);
+  position: relative;
+}
+/* DQ ウィンドウの四隅装飾 (dq-window と同じ) */
+.workspace-column::before,
+.workspace-column::after {
+  content: '';
+  position: absolute;
+  width: 6px;
+  height: 6px;
+  background: var(--color-border);
+  pointer-events: none;
+}
+.workspace-column::before { top: 3px; left: 3px; }
+.workspace-column::after  { bottom: 3px; right: 3px; }
+
+.workspace-column-header {
+  flex: 0 0 auto;
+  padding: 0.45em 0.8em;
+  font-size: 0.92em;
+  font-weight: 700;
+  border-bottom: 2px solid var(--color-window-inner-border);
+}
+
+.workspace-column-body {
+  flex: 1 1 auto;
+  overflow-y: auto;
+  overscroll-behavior: contain;   /* 縦終端が外側に漏れない */
+  padding: 0.6em 0.7em;
+}
+
+@media (min-width: 768px) {
+  .workspace-column {
+    flex: 0 0 340px;
+    /* header + footer-nav 分を引いた高さ。実測 CSS 変数化は PR 5 で。 */
+    height: calc(100dvh - 9.5em);
+  }
+}
+
+@media (max-width: 767.98px) {
+  .workspace-column {
+    /* モバイル縦積み: 1 カラムが画面を占有しすぎない高さに */
+    max-height: calc(100dvh - 12em);
+  }
+}
+
 .header {
   padding: 0.6em 1em;
   font-size: 1.05em;
